requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
openpyxl
python-dotenv

# Django FHIR Platform
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery
google-cloud-storage
google-api-python-client
